move array comparison util
