+++ packages/core/src/library/CommonLib.ts
@@ -23,6 +23,30 @@ export const arraysAreEqual = <T>(array1: T[], array2: T[]): boolean => {
 };
 
 /**
+ * @name areSortedArraysEqual
+ * @summary Checks if two sorted string arrays are equal, returns `true` or `false`.
+ */
+export const areSortedArraysEqual = (
+  arr1: string[],
+  arr2: string[]
+): boolean => {
+  if (arr1.length !== arr2.length) {
+    return false;
+  }
+
+  const sortedArr1 = [...arr1].sort();
+  const sortedArr2 = [...arr2].sort();
+
+  for (let i = 0; i < sortedArr1.length; i++) {
+    if (sortedArr1[i] !== sortedArr2[i]) {
+      return false;
+    }
+  }
+
+  return true;
+};
+
+/**
  * @name generateUID
  * @summary Util for generating a UID in the browser.
  */
